requirement.txt
streamlit
pypdf
sentence-transformers
faiss-cpu
numpy